# openapi.json
{
  "components": {
    "schemas": {
...
      "ConfigBacktest": {
        "additionalProperties": false,
        "properties": {
          "commission_pct": {
            "title": "Commission Pct",
            "type": "number"
          },
          "initial_capital": {
            "title": "Initial Capital",
            "type": "number"
          },
          "lookback_days": {
            "title": "Lookback Days",
            "type": "integer"
          },
          "slippage_pct": {
            "title": "Slippage Pct",
            "type": "number"
          },
          "stop_loss_pct": {
            "title": "Stop Loss Pct",
            "type": "number"
          },
          "take_profit_pct": {
            "title": "Take Profit Pct",
            "type": "number"
          }
        },
        "required": [
          "lookback_days",
          "initial_capital",
          "commission_pct",
          "slippage_pct",
          "stop_loss_pct",
          "take_profit_pct"
        ],
        "title": "ConfigBacktest",
        "type": "object"
      },
      "ConfigML": {
        "additionalProperties": false,
        "properties": {
          "lookback_bars": {
            "title": "Lookback Bars",
            "type": "integer"
          },
          "min_confidence": {
            "title": "Min Confidence",
            "type": "number"
          },
          "n_estimators": {
            "title": "N Estimators",
            "type": "integer"
          },
          "train_test_split": {
            "title": "Train Test Split",
            "type": "number"
          }
        },
        "required": [
          "train_test_split",
          "n_estimators",
          "lookback_bars",
          "min_confidence"
        ],
        "title": "ConfigML",
        "type": "object"
      },
      "ConfigResponse": {
        "additionalProperties": false,
        "properties": {
          "backtest": {
            "$ref": "#/components/schemas/ConfigBacktest"
          },
          "ml": {
            "$ref": "#/components/schemas/ConfigML"
          },
          "pairs": {
            "items": {
              "$ref": "#/components/schemas/PairRef"
            },
            "title": "Pairs",
            "type": "array"
          },
          "risk": {
            "$ref": "#/components/schemas/ConfigRisk"
          },
          "validation": {
            "$ref": "#/components/schemas/ConfigValidation"
          },
          "watchlist": {
            "additionalProperties": true,
            "title": "Watchlist",
            "type": "object"
          }
        },
        "required": [
          "watchlist",
          "pairs",
          "backtest",
          "validation",
          "risk",
          "ml"
        ],
        "title": "ConfigResponse",
        "type": "object"
      },
      "ConfigRisk": {
        "additionalProperties": false,
        "properties": {
          "max_correlated_positions": {
            "title": "Max Correlated Positions",
            "type": "integer"
          },
          "max_drawdown_halt_pct": {
            "title": "Max Drawdown Halt Pct",
            "type": "number"
          },
          "max_portfolio_risk_pct": {
            "title": "Max Portfolio Risk Pct",
            "type": "number"
          },
          "max_position_pct": {
            "title": "Max Position Pct",
            "type": "number"
          }
        },
        "required": [
          "max_position_pct",
          "max_portfolio_risk_pct",
          "max_correlated_positions",
          "max_drawdown_halt_pct"
        ],
        "title": "ConfigRisk",
        "type": "object"
      },
      "ConfigValidation": {
        "additionalProperties": false,
        "properties": {
          "min_profit_factor": {
            "title": "Min Profit Factor",
            "type": "number"
          },
          "min_sharpe": {
            "title": "Min Sharpe",
            "type": "number"
          },
          "min_trades": {
            "title": "Min Trades",
            "type": "integer"
          },
          "min_win_rate": {
            "title": "Min Win Rate",
            "type": "number"
          },
          "signal_recency_days": {
            "title": "Signal Recency Days",
            "type": "integer"
          }
        },
        "required": [
          "min_win_rate",
          "min_profit_factor",
          "min_sharpe",
          "min_trades",
          "signal_recency_days"
        ],
        "title": "ConfigValidation",
        "type": "object"
      },
      "DataStatusResponse": {
        "properties": {
          "symbols": {
            "items": {
              "$ref": "#/components/schemas/StoreSymbolStatus"
            },
            "title": "Symbols",
            "type": "array"
          }
        },
        "required": [
          "symbols"
        ],
        "title": "DataStatusResponse",
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "description": "The ``error`` object of the envelope (documented in OpenAPI for every non-2xx response).",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "Body of every non-2xx response: ``{\"error\": {\"code\", \"message\", \"details\", \"request_id\"}}``.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "PairRef": {
        "additionalProperties": false,
        "properties": {
          "a": {
            "title": "A",
            "type": "string"
          },
          "b": {
            "title": "B",
            "type": "string"
          }
        },
        "required": [
          "a",
          "b"
        ],
        "title": "PairRef",
        "type": "object"
      },
...
      "StoreSymbolStatus": {
        "additionalProperties": false,
        "description": "SymbolStatus plus the bar store's stale flag.",
        "properties": {
          "adjusted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adjusted"
          },
          "cache_age_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Age Seconds"
          },
          "calendar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calendar"
          },
          "coverage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "fetched_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fetched At"
          },
          "last_session": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Session"
          },
          "n_bars": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "N Bars"
          },
          "quality_flags": {
            "items": {
              "type": "string"
            },
            "title": "Quality Flags",
            "type": "array"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "stale": {
            "default": false,
            "title": "Stale",
            "type": "boolean"
          },
          "stale_sessions": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stale Sessions"
          },
          "status": {
            "default": "ok",
            "title": "Status",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol"
        ],
        "title": "StoreSymbolStatus",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Trading Bot v2 API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/config/": {
      "get": {
        "description": "Get current configuration.",
        "operationId": "get_config_api_config__get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthorized (API_TOKEN set, token missing or wrong)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unknown_symbol / no_data / not_found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error / unknown_pattern / insufficient_data"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "internal_error / contract_violation"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "data_quality"
          }
        },
        "summary": "Get Config",
        "tags": [
          "config"
        ]
      }
    },
    "/api/data/status": {
      "get": {
        "description": "Per-symbol data source, coverage, staleness and quality (may fetch bars, so it shares the heavy-endpoint cap: 429 when busy; default: watchlist + allocation).",
        "operationId": "get_data_status_api_data_status_get",
        "parameters": [
          {
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Za-z0-9.\\-=^]{1,20}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthorized (API_TOKEN set, token missing or wrong)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unknown_symbol / no_data / not_found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error / unknown_pattern / insufficient_data"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "busy: another heavy request is running"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "internal_error / contract_violation"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "data_quality"
          }
        },
        "summary": "Get Data Status",
        "tags": [
          "data"
        ]
      }
    },
...
  }
}